setup with intellij
