+++ javainterviewapp/src/main/java/com/javainterview/app/stringInterview/AnagramImpl.java
@@ -1,0 +1,18 @@
+package com.javainterview.app.stringInterview;
+
+public class AnagramImpl implements Anagram<Anagram>  {
+    public String value;
+
+    public AnagramImpl(String s) {
+        this.value = s;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public int compareTo(Anagram o) {
+        return value.compareTo(o.getValue());
+    }
+    
+}
